feat: Tiempo con Decimales
- Con formato 0:00:000 (minutos, segundos, milisegundos).

+++ Assets/Scripts/Save/DataGame.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 [Serializable]
 public class DataGame
 {
-    public List<ushort> _pointsPerLevel = new List<ushort>();
+    public List<float> _pointsPerLevel = new List<float>();
 }
